+++ src/V1/Exceptions/DefensiveExceptions.php
@@ -57,6 +57,8 @@ class DefensiveExceptions extends FactoryListContainer
             'BadAssuranceArgs::newFromVar' => [ BadAssuranceArgs::class, 'newFromVar' ],
             'BadAssurancesList::newFromInputParameter' => [ BadAssurancesList::class, 'newFromInputParameter' ],
             'BadAssurancesList::newFromVar' => [ BadAssurancesList::class, 'newFromVar' ],
+            'BadCallable::newFromInputParameter' => [ BadCallable::class, 'newFromInputParameter' ],
+            'BadCallable::newFromVar' => [ BadCallable::class, 'newFromVar' ],
             'EmptyAssurancesList::newFromInputParameter' => [ EmptyAssurancesList::class, 'newFromInputParameter' ],
             'EmptyAssurancesList::newFromVar' => [ EmptyAssurancesList::class, 'newFromVar' ],
             'BadRequirement::newFromInputParameter' => [ BadRequirement::class, 'newFromInputParameter' ],